extract command argument objects
