+++ node-runner-cli/commands/ledgercommand.py
@@ -1,11 +1,10 @@
 from argparse import ArgumentParser
 from argparse import RawTextHelpFormatter
 
+import boto3
 import botocore
 
 from commands.subcommand import get_decorator, argument
-
-import boto3
 
 ledgercli = ArgumentParser(
     description='Subcommand to help to sync up the ledger from a S3 bucket',
